requirements.txt
asn1crypto==0.23.0
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.4.8
botocore==1.8.6
cement==2.8.2
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
colorama==0.3.7
coverage==4.4.2
cryptography==2.1.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.32.0
django-autoslug-field==0.2.3
django-ckeditor==5.3.0
django-comments-xtd==2.0.8
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-extensions==1.9.6
django-hitcount==1.2.3
django-js-asset==0.1.1
django-newsletter==0.6
django-storages==1.6.5
django-tastypie==0.14.0
django-webmaster-verification==0.3.0
django-webtest==1.9.2
djangorestframework==3.6.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
ldif3==3.1.1
lxml==4.1.1
mysqlclient==1.3.12
oauthlib==2.0.2
olefile==0.44
pathspec==0.5.0
Pillow==4.2.1
psycopg2==2.7.1
pycparser==2.18
python-card-me==0.9.3
python-dateutil==2.6.1
python-decouple==3.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.12
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.3
stripe==1.62.1
surlex==0.2.0
tabulate==0.7.5
termcolor==1.1.0
unicodecsv==0.14.1
urllib3==1.22
waitress==1.1.0
WebOb==1.7.3
websocket-client==0.44.0
WebTest==2.0.29
